fix bug04

+++ src/main/java/com/amigoscode/bughunt/easy/bug04/Percentage.java
@@ -3,6 +3,6 @@ package com.amigoscode.bughunt.easy.bug04;
 public class Percentage {
 
     public double percentage(int part, int total) {
-        return (part / total) * 100;
+        return ((double)part / total) * 100;
     }
 }
